# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CountResponse:
    properties:
      count:
        type: integer
    type: object
...
  dto.CreateProdutoRequest:
    properties:
      ativo:
        description: pointer para permitir false explícito
        type: boolean
      categoria:
        maxLength: 100
        type: string
      descricao:
        maxLength: 1000
        type: string
      estoque:
        minimum: 0
        type: integer
      nome:
        maxLength: 200
        minLength: 3
        type: string
      preco:
        type: number
      sku:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - nome
    - preco
    - sku
    type: object
  dto.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  dto.ProdutoResponse:
    properties:
      ativo:
        type: boolean
      categoria:
        type: string
      created_at:
        type: string
      descricao:
        type: string
      estoque:
        type: integer
      id:
        type: integer
      nome:
        type: string
      preco:
        type: number
      sku:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.UpdateProdutoRequest:
    properties:
      ativo:
        type: boolean
      categoria:
        maxLength: 100
        type: string
      descricao:
        maxLength: 1000
        type: string
      estoque:
        minimum: 0
        type: integer
      nome:
        maxLength: 200
        minLength: 3
        type: string
      preco:
        type: number
      sku:
        maxLength: 50
        minLength: 3
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@exemplo.com
    name: API Support
  description: REST API for managing clientes (customers) with CRUD operations
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Cliente API
  version: "1.0"
paths:
...
  /produtos:
    get:
      description: Retrieve all produtos from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ProdutoResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get all produtos
      tags:
      - produtos
    post:
      consumes:
      - application/json
      description: Create a new produto with the provided information
      parameters:
      - description: Produto data
        in: body
        name: produto
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProdutoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ProdutoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a new produto
      tags:
      - produtos
  /produtos/{id}:
    delete:
      description: Delete a produto by ID
      parameters:
      - description: Produto ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete produto
      tags:
      - produtos
    get:
      description: Retrieve a specific produto by ID
      parameters:
      - description: Produto ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProdutoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get produto by ID
      tags:
      - produtos
    put:
      consumes:
      - application/json
      description: Update an existing produto
      parameters:
      - description: Produto ID
        in: path
        name: id
        required: true
        type: integer
      - description: Produto data
        in: body
        name: produto
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProdutoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProdutoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update produto
      tags:
      - produtos
  /produtos/categoria/{categoria}:
    get:
      description: Retrieve produtos by categoria
      parameters:
      - description: Produto categoria
        in: path
        name: categoria
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ProdutoResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get produtos by categoria
      tags:
      - produtos
  /produtos/count:
    get:
      description: Get the total number of produtos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CountResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Count produtos
      tags:
      - produtos
  /produtos/nome/{name}:
    get:
      description: Retrieve produtos matching the specified name (partial match)
      parameters:
      - description: Produto name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ProdutoResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get produtos by name
      tags:
      - produtos
swagger: "2.0"